+++ programs/solot/Cargo.toml
@@ -18,5 +18,4 @@ default = []
 [dependencies]
 anchor-lang = { version = "0.29.0", features = ["init-if-needed"] }
 anchor-spl = {version = "0.29.0",features = ["metadata"]}
-mpl-token-metadata = { version="4.0.0", features = [  ] }
 solana-program = "1.18.8"
