+++ programs/solot/src/constants.rs
@@ -1,0 +1,5 @@
+use solana_program::{pubkey, pubkey::Pubkey};
+
+pub const ADMIN_PUBKEY: Pubkey = pubkey!("D2pRGmGrgkUKWvv5KXUcGo7R7bGtg6m6yqSYr2x29fZA");
+
+pub const INIT_SOLOT: u64 = 10000;
