+++ programs/solot/src/instructions/create_mint.rs
@@ -1,0 +1,79 @@
+use crate::constants::ADMIN_PUBKEY;
+use anchor_lang::prelude::*;
+use anchor_spl::metadata::mpl_token_metadata::accounts::Metadata as MetadataAccount;
+use anchor_spl::metadata::mpl_token_metadata::types::DataV2;
+use anchor_spl::metadata::{create_metadata_accounts_v3, CreateMetadataAccountsV3, Metadata};
+use anchor_spl::token::{Mint, Token};
+
+#[derive(Accounts)]
+pub struct CreateMint<'info> {
+    #[account(mut, address = ADMIN_PUBKEY)]
+    pub admin: Signer<'info>,
+    #[account(
+        init,
+        seeds = [b"reward"],
+        bump,
+        payer = admin,
+        mint::decimals = 9,
+        mint::authority = token_mint,
+
+    )]
+    pub token_mint: Account<'info, Mint>,
+
+    ///CHECK: Using "address" constraint to validate metadata account address, this account is created via CPI in the instruction
+    #[account(
+        mut,
+        address = MetadataAccount::find_pda(&token_mint.key()).0,
+    )]
+    pub metadata_account: UncheckedAccount<'info>,
+
+    pub token_program: Program<'info, Token>,
+    pub token_metadata_program: Program<'info, Metadata>,
+    pub system_program: Program<'info, System>,
+    pub rent: Sysvar<'info, Rent>,
+}
+
+impl<'info> CreateMint<'info> {
+    pub fn handler(
+        ctx: Context<CreateMint>,
+        uri: String,
+        name: String,
+        symbol: String,
+    ) -> Result<()> {
+        let seeds = b"reward";
+        let bump = ctx.bumps.token_mint;
+        let signer: &[&[&[u8]]] = &[&[seeds, &[bump]]];
+        // CPI Context
+        let cpi_ctx = CpiContext::new_with_signer(
+            ctx.accounts.token_metadata_program.to_account_info(),
+            CreateMetadataAccountsV3 {
+                metadata: ctx.accounts.metadata_account.to_account_info(), // the metadata account being created
+                mint: ctx.accounts.token_mint.to_account_info(), // the mint account of the metadata account
+                mint_authority: ctx.accounts.token_mint.to_account_info(), // the mint authority of the mint account
+                update_authority: ctx.accounts.token_mint.to_account_info(), // the update authority of the metadata account
+                payer: ctx.accounts.admin.to_account_info(), // the payer for creating the metadata account
+                system_program: ctx.accounts.system_program.to_account_info(), // the system program account, required when creating new accounts
+                rent: ctx.accounts.rent.to_account_info(), // the rent sysvar account
+            },
+            signer, // pda signer
+        );
+        let data_v2 = DataV2 {
+            name: name,
+            symbol: symbol,
+            uri: uri,
+            seller_fee_basis_points: 0,
+            creators: None,
+            collection: None,
+            uses: None,
+        };
+
+        create_metadata_accounts_v3(
+            cpi_ctx, // cpi context
+            data_v2, // token metadata
+            true,    // is_mutable
+            true,    // update_authority_is_signer
+            None,    // collection details
+        )?;
+        Ok(())
+    }
+}
