+++ programs/solot/src/instructions/mod.rs
@@ -1,0 +1,5 @@
+pub mod create_mint;
+pub mod initialize_lottery;
+
+pub use create_mint::*;
+pub use initialize_lottery::*;
